Remove deprecated backend files and update API base URL in frontend extension
