+++ backend/src/architecture.png
